+++ api/src/app.js
@@ -3,11 +3,12 @@ const cors = require('cors')
 const app = express()
 
 const { getSingleWordHandler } = require('./handlers/words')
-const { getImageHandler } = require('./handlers/banner')
 const { getSummaryHandler } = require('./handlers/summary')
 const { getSuggestions } = require('./handlers/suggestions')
 
 const getWordsOfDict = require('./dict/getWordsOfDict')
+
+const getBanner = require('./banner/getBanner')
 
 const { validateGetSuggestions } = require('./validators/suggestions')
 
@@ -33,7 +34,7 @@ app.get('/words/:word', getSingleWordHandler)
 
 app.get('/dict/:name', getWordsOfDict.validator, getWordsOfDict.handler)
 
-app.get('/banner/:word', getImageHandler)
+app.get('/banner/:word', getBanner.handler)
 
 app.get('/summary', getSummaryHandler)
